Fixing margin of CalendarPicker control
Spotlight strangely behave when using the five way to navigate between
days on the calendar. This is because of each day control have not the
same margin width when the margin is set 'auto'.

To fix this, revert "margin: 0 auto;" to "margin: 0 2.26%;" in
.moon-calendar-picker-date.

Enyo-DCO-1.1-Signed-off-by: Kunmyon Choi kunmyon.choi@lge.com