defect fix for GF-12211
Defect reason: The "client" component  was exceeding beyond the
boundaries.
As all the other components were moved to top, to provide proper view
for the drawer, however the "client" component was not been moved and as
per the styles it was taking the height of the parent div.
Solution: provided top to "client"component also.

Enyo-DCO-1.1-Signed-off-by: Anish Ramesan <anish.ramesan@lge.com>