Defect GF-12213 fix
As the per the defect the scrolling should happen for one single page at
a time.
However the "paginationPageMultiplier" is defaulted to 10, which lead to
scrolling of 10 rows at a time.
To fix this issue, I think we could paginate as per the scroll bound
rather than the scroll delta(which always calculates to a 187.5). And
this way will also justify for the name of the published property
pagination-page-multiplier.
Please provide your suggestion.

Enyo-DCO-1.1-Signed-off-by: Anish Ramesan anish.ramesan@lge.com