Bug GF-9875 fix
Defect is related to nub positioning.
Adjusted the nub position.
Set proper image for the nub when on left side.
Generated moonstone.css

There are two more changes in Moonstone.css,, as some of changes made in
less file where not applied on the generated moonstone.css. Including
that change also in this fix.