Bug GF-11831 fix
As per the feedback, I have made changes in ContextualPopup and Popup
components.
Now these components will have moon.IconButton as close button.
Removed the css properties which will not be required as those
properties are already set in moon.IconButton.

This commit cannot be merged as this requires an asset(close icon).