GF-30500 marquee flicker defect fix
Issue:
After the marquee animation ends there is flicker before its holds on to
the current state. Observed in TV only.
Analysis:
Webit "marquee" animation is used for text scrolling animation. However
when the animation ends, as per the default behavior of an animation, it
moves to the starting frame. In this case also, when the animation ends
the frame moves to the starting position, but we have asked to hold for
some duration at last frame. So this results in a flicker, as frames are
moved from start to end.

Enyo-DCO-1.1-Signed-off-by: Anish Ramesan <anish.ramesan@lge.com>