GF-33510 HeaderAutoCollapsingSample Update
If multiple panels are visible in viewport area,by default 0(zero) is
the active panel.hence if 2 panels are visible, to move to the 3rd
panel, 2 times next() has to be called(it waits for 2 clicks).

Solution : As in other pattern samples setIndex() method can be directly
used instead of next.