GF-33990-on resizing drawers scrollHeight is incorrect
While resizing, the controlDrawer ScrollHeight is not calculated
correctly in moon.Drawer. This causes the drawer to draw some extra
height on resizing.

onresize handler, controlDrawer is rendered, which makes the improper
alignment of the client within it. This causes the controlDrawer to have
an incorrect ScrollHeight. So to fix the bug, the client is resized
after the controlDrawer is rendered.

Enyo-DCO-1.1-Signed-off-by: Rajyavardhan P <rajyavardhan.p@lge.com>