GF-35019 SimplePicker issue when parent not shown
Issue: When parent component of SimplePicker is hidden(set display
none), and then shown again, SimplePicker inner components are getting
not proper bounds. This is due to the fact that bounds of inner divs
will not be proper when the parent div is invisible. So in this case
when the parent is again made visible, the child divs have to be
rendered again.
Solution: As SimplePicker depends on the visibility of parent component,
so added a check for that.

Updated the sample to have this implementation.