GF-36646-Defect Fix
height property missing for title above when No content is provided.
Since the height is calculated based on the content .Height property
will be set to Zero(0) When content is empty.(Hence overlaps)
Solution : when content is not there, CSS Property is already exists for
Removing Border, Adding height property to that solves this problem.
Enyo-DCO-1.1-Signed-off-by: Ashwini R <ashwini13.r@lge.com>