Avoid scrolling both vertically and horizontally with wheel
When both horizonal and vertical scroll bars are on screen,
the code would try adjusting the postiion of both if a user
moved the scroll wheel.  This caused diagonal movement and
clearly wasn't intented. This changes it to give vertical
movement the preference, although on mice with both direction
scrolling, both x and y will be handled.

Enyo-DCO-1.1-Signed-Off-By: Ben Combee (ben.combee@lge.com)