GF-36142 SmallHeader on RTL Header Overlap Defect Fixed
Issue: Header: Overlapped Header Occurs When Toggle RTL is Off.
When RTL is ON 'padding-left' was given to SmallHeader, but when RTL is
OFF padding-left was present which is not needed and no padding-right
was there.

Fixes:
Added class '.enyo-locale-right-to-left .moon-small-header
.moon-header-title-wrapper' to  'Header.less' with "padding-right"
property and padding-left made 0.

Enyo-DCO-1.1-Signed-off-by: Anugrah Saxena <anugrah.saxena@lge.com>