GF-6488-defect fixed- updated the TransitionInfos
photo slide control: when click on the picture for slide show, it lands
on the 2nd picture.

Cause: The application uses flyweight pattern to render the slideshow
images on to the page. But in panels the transitionInfo values are
unchanged, which is blocking the transitionFinish and transitionStart
events from triggering. So this makes the slideshow to stop at the 2nd
picture.
Fix: The startTransitionInfo and finishTransitionInfo values are updated
in accordance with the panels for every iteration  to fix the bug.

Enyo-DCO-1.1-Signed-off-by: Rajyavardhan P <rajyavardhan.p@lge.com>