GF-35777: don't create layers for ExpandableListItems
Creating separate layers for the expandable list items is hurting full screen
performance and leading to OOM errors on EnyoBench.  Removing this acceleration
doesn't seem to affect on-screen perf very much since the drawers are pushing
other content, leading to layout changes and repaints, so we'll disable this until
we can find a better scheme for drawers that expand in-line.

Enyo-DCO-1.1-Signed-Off-By: Ben Combee (ben.combee@lge.com)