GF-41734 Defect Fixed VideoPlayer when FastForward reaching Buffer End
Issue: moon.VideoPlayer: When fast-forward reaches end of buffer, switch
back to play.
To handle this case it was needed to use a HTML5 video loop event in
which buffered content and current content can be monitored.
Fixes:
Added case in timeupdate to handle fast-forward when it reaches end of
buffer, and added API 'onFFBufferEnd' to switch video to play mode and
update playback control to work as expected.

Enyo-DCO-1.1-Signed-off-by: Anugrah Saxena <anugrah.saxena@lge.com>