GF-41734 Defect Fixed VideoPlayer when FastForward reaching Buffer End
Issue: moon.VideoPlayer: When fast-forward reaches end of buffer, switch
back to play.
No event is triggered by video to tell that video has reached the end of
buffer, and there is no way to know in which buffer range the current
play is in to give end condition for multiple buffer ranges case.
Fixes:
Added condition for fast-forward in timeupdate to change playback to
play mode when video reaches end of
buffer.
Added "getBufferCurrentRangeIndex" to check in which buffer range our
playback is currently.

Enyo-DCO-1.1-Signed-off-by: Anugrah Saxena <anugrah.saxena@lge.com>