Modify ExpandablePicker sampler to have auto mode
This is activated by adding #auto to the URL used to load the sample,
it will go through all the pickers opening and closing the over and
over forever. This is for testing WebKit to see if we're causing any
rendering memory leaks.

Enyo-DCO-1.1-Signed-Off-By: Ben Combee (ben.combee@lge.com)