Fixing http://jira2.lgsvl.com/browse/GF-42447
moon.Popup should not give focus to first child when it is opened in
pointer mode.

The container decorator is already have defaultSpotlightControl property
and popup itself is container so we don't need to add defaultControl.

And moon.Popup has _spotlight private property which is caching last
spotlight property but it is conflict with container decorator in
spotlight so I modified _spotlight to _spotlightOld.

Enyo-DCO-1.1-Signed-off-by: Kunmyon Choi (kunmyon.choi@lge.com)