GF-43095-PanelsVideoPlayerSample-Inline Viedo Overlapping Issue
This is the sample issue not the framework issue, in the sample
moon.Item "Capture" is initially aligned to right. when we turn on the
RTL, the alignment of this item is not reflected as desired, that is why
'Capture' item is overlapping on the Inline Video.
so added a class to the sample.css which can do the job when RTL is
turned on.

Enyo-DCO-1.1-Signed-off-by: Srinivas V <srinivas.v@lge.com>