GF-38531 SimpleIntegerPicker: Broken Picker Displays
When simple integerpicker is not focused and just change index in
programmatic way, there is a case that pink half circle button is
showing while it is in animation.

To guarantee this behavior is not happens, adding animation class which
is doing the same thing as spotlight. And add remove class while it is
in the middle of animation.

DCO-1.1-Signed-Off-By: Kunmyon Choi <kunmyon.choi@lge.com>