GF-52877-InputHeaders-Truncated Placeholders Display-and-Hidden titleAbove Issue
When we turn non-latin on, due to the wrong width value of moon.input
field:  placeholders in InputHeaders looks truncated. so adjusted the
width value of moon.input field when non-latin turned on. and for the
Non-reported issue: In the Small Input header 'titleAbove' tag is not
visible, because when we set "smallHeader" true the height of
'titleAbove" is setted as 0, so fixed this issue too in this commit.

Enyo-DCO-1.1-Signed-off-by: Srinivas V
<srinivas.v@lge.com>