GF-53518-defect Fixed
DatePicker : request exception handling in minYear and maxYear

If "value" property is not used in any subkind the picker shows the
current date as the default value irrespective of minYear and maxYear in
the context.
So to handle the exception getPickerDate() method is introduced to
return the "value" property based on the minYear and maxYear.

Enyo-DCO-1.1-Signed-off-by: RajyavardhanP <rajyavardhan.p@lge.com>