GF-60073: Remaining focus when datalist is refreshing
There are some cases that data item is having more then one focusable
control.

In this case, the focus will be given to the first child of the item
after refreshing data list.

This fix is checking the previously focused item is not generated and it
is give focus to the same control if it is not removed. And it is try to
give focus to the last item if previous focused item is removed.

By doing this way, focus will be remains in the same control position.

DCO-1.1-Signed-Off-By: Kunmyon Choi <kunmyon.choi@lge.com>