Update WIP moon.AudioPlayback kind, add sample
moon.AudioPlayback is broken because of the removal of moon.List.
This doesn't fix that, but it adds notes indicating that it's broken
and adds back a sample that can be used to test it when the control
is fixed in the future.

Enyo-DCO-1.1-Signed-Off-By: Ben Combee (ben.combee@lge.com)