In moon.Popup, don't allow focus to escape via 5-way moves (BHV-8502).
We do this by adding onSpotlightFocus to the set of events that the popup captures when it's open. We inspect the captured event and decide (using the public Spotlight APIs for last and current focus) whether to allow focus to leave (and close the popup) or keep focus inside.

This change required us to make eventsToCapture a concat'able property in enyo.Popup (implemented in a separate pull request to the enyo repo).

Making this change also allowed us to no longer rely on the onSpotlightContainerLeave event to close the popup, which in turn allowed us to get rid of the "spotlightDummy" DOM node formerly required for cases where nothing in the popup was spottable.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)