BHV-10787: Fix 5-way show/hide of video player controls.
This fix appears to address the issue, so we should take it for now (assuming there are no issues discovered in integration), but the root cause is still somewhat mysterious. The spotSentFromContainer functionality added to Spotlight and formerly used here to properly distinguish up / down events that the player should handle seems to be working differently as compared to Enyo 2.3. For now, I've revised the VideoPlayer logic to avoid using `spotSentFromContainer`, but we should go back and diagnose what's going on with that feature.

Also made a handful of changes to address other issues with 5-way navigating VideoPlayer that I bumped into while testing against PanelsVideoPlayerSample and AlwaysViewingPanelsWithVideoSample.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)