BHV-11579: Fixing ListActions focus lost when ListAction drawer is closed.
### Analysis
In 2.3.0, listActions drawer was showing false when it is closed.
But, now it is showing true and this makes the controls inside of drawer can be a candidate of getNearestPointerNeighbor algorithm.

### Solution
Makes spotlightDisabled which is having relatively low cost.

In addition to this, drawer client is off screen when it is closed. The drawer itself is placed on the header area.
This is depends on the creation time of activator control, but sometimes this makes pointer event not reached to the activator.
So, we can apply pointer-mode as none on only the drawer.

DCO-1.1-Signed-Off-By: Kunmyon Choi kunmyon.choi@lge.com