ANG-11530: Adding direction ltr on IntegerPicker and Revert ilib format from SimpleIntegerPicker
This is a requirement from corporate lab for Farsi/Arabic.

### Issue:
Moonstone controls should not change format on negative numeric values.

### Fix:
Adding direction ltr on IntegerPicker and SimpleIntegerPicker so that it is not affected by enyo-locale-right-to-left.
Revert ilib format fix from SimpleIntergerPicker so that is is never change format.

DCO-1.1-Signed-Off-By: Kunmyon Choi kunmyon.choi@lge.com