BHV-11417: Moonstone Component Bug for ExpandableIntegerPicker - step
### Issue
ExpandableIntegerPicker have SimpleIntegerPicker inside.
When step is set to non valid value, all arrow buttons are hidden.
Because, "is" is equals to "min" and "max" in the setButtonVisibility below.
But, when step is set to 1 again, "was" is null and "is" is 1.
This doesn't matches any visible case. Because, all the visible case is looking at "was" not "is".
And, when set step as non-integer value the it is enters to infinite loop.

### Fix
Add condition to check type on was and step.

DCO-1.1-Signed-Off-By: Kunmyon Choi kunmyon.choi@lge.com