BHV-14435: Updating drawer activator during create
"icon" and "src" are not updated for activator when "drawers" component
block is not present.
When drawer components are not present, the "onDrawersRendered" event
which is waterfall'd during "moon.Drawers" render, is not received by
defaulted "moon.Drawer" component. Thus not leading to trigger
"updateActivator" function required to set the "icon" and "src".