BHV-14466: Panels overlapped when push listAction panel with 5 way 'OK' key press
Issue:

Panel doesn't shrink to breadcrumb while panel arrangement when press 'ok' on panel item to push panel which is having listAction.
ListAction is close itself after completing render, and give focus to activator when it is close itself.
Panel have a code like when item on panel getting focused it is change index.
When panel change index, it is query transition info (getTransitionInfo) which is calling isBreadcrumb.
In this step, arranger is never calculated breadcrumbPositions values for second panel because it is before arrange.

Fix:

Add option on oncomplete event like "skipSpot: true" to let drawerAnimationEnd handler know it is called on render.
Adding listaction on sample for testing this issue.

DCO-1.1-Signed-Off-By: Kunmyon Choi kunmyon.choi@lge.com