CMB-1725: moon.VideoPlayer: Add handling of pointer hidden/shown special keyIdentifiers for hiding/showing OSD
Issue:
Currently, moon.VideoPlayer uses the "autoCloseTimeout" property to hide the overlay after mousemoves stop. The "shakeAndWake" mode also uses mousemove to detect when the pointer became active.
We should keep the autoCloseTimeout for cross-platform use, but also implement key handling of the special "pointer hidden/shown" key events that are dispatched when the magic remote pointer is hidden/shown. If the VideoPlayer receives the "shown" keyIdentifier and "shakeAndWake" is true, it should activate the overlay. If the VideoPlayer receives the "hidden" keyIdentifier, it should cancel the autoCloseTimeout and hide the overlay immediately.
Please also test on the TV, to ensure that the default autoCloseTimeout is slightly longer than the default magic remote timeout.

Fix:
Add a logic that checking keyCode on remoteKeyHandler of VideoPlayer to show/hide OSD.

DCO-1.1-Signed-Off-By: Kunmyon Choi kunmyon.choi@lge.com