ENYO-483: Replace calls to stabilize() with explicit scroll events
A recent launch time regression for the SmartShare app was traced
back to extraneous layouts in the app, which were caused by
unnecessary onScroll events resulting from calls to
ScrollMath.stabilize(). In Enyo core, we reworked stabilize() to
eliminate the unnecessary scroll events. However,
moon.ScrollStrategy was relying on stabilize() generating scroll
events to implement the hold-to-repeat behavior of its paging
controls, so in this PR we replace the calls to stabilize() with
explicit calls to generate the required scroll events.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)