hide overflow of checkbox items
Without overflow: hidden, when a CheckboxItem is included in a ListAction list, it can be spotted by 5-way but may cause the paging controls to be incorrect. This occurs when spotting the last item in the list because it is scrolled into view but the overflowing tap area does not completely come into the viewport causing the scroller to report there is still more to scroll.

Issue: BHV-17852
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)