ENYO-544: Restore manual call to stabilize()
In enyo.ScrollMath, for ENYO-483, we introduced a call to
stabilize() every time setScrollX() or setScrollY() was called.
This seemed like a logical place to clamp the scroll position and
eliminated the need for some manual calls to stabilize() elsewhere,
but it turns out that some existing behavior in moon.ScrollStrategy
depends on allowing the position to be temporarily out of bounds.

The Moonstone behavior is complicated, but seems to be reasonable,
so reverting this particular change (affecting ScrollMath,
TouchScrollStrategy and TranslateScrollStrategy in Enyo, as well as
moon.ScrollStrategy.)

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)