replace paging throttle with custom impl
We want the paging controls update immediately when a scroll starts so if the user scrolls from the top, the top arrow will be disabled immediately. We also want to throttle future calls during a scroll because the status doesn't change that frequently. The job API doesn't support this requirement so I've added a custom implementation.

Issue: ENYO-530
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)