BHV-18024: Display negative sign properly for locale
moon.Slider was only applying locale-specific formatting to its
value when showPercentage was true. Consequently, the placement
of the negative sign for negative numbers (when showPercentage was
false, as it normally would be for a slider supporting negative
values) was not locale-sensitive. This caused problems because
the text direction of the slider's value popup was properly set
per locale, but the value string was not, resulting in the negative
sign being displayed on the wrong side for some locales.

To address this issue, we now apply locale-specific formatting of
the popup value in all cases.

Also modifying the Slider sample to include a slider with a
negative range, making this type of issue easier to test going
forward. (Credit to Aaron Tam.)

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)