add hd mixin to convert box model properties to safe values
Since the base size is 1080p, values converted to 720p are multiplied
by 2/3 which results in sub-pixel values for many properties. The
effect is that the browser may alter the dimensions of a node based
on the accumulation of the fractional part of box model property
values.

This mixin rounds the value to the nearest multiple of 3 so the
conversion results in a round integer value thereby avoiding the
fractional accumulation issue

Issue: ENYO-736
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)