refactor picker client positioning logic into new method
By calling selectedIndexChanged during rendered(), SimplePicker was
firing its onChange event unnecessarily because it's guarded by
a check on this.generated. Since the purpose of calling it was to
adjust the position of the client based on the calculated size of the
picker controls, we can refactor that into another method and call
that instead of selectedIndexChanged.

Issue: ENYO-800
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)