ENYO-792: Fix marquee on moon.ItemOverlay
Issue:
When moon.Item has marqueeText inside of components block, it is creating div inside of moon.item.
In this case, the moon.Item has no content in this.content. But, the content is on marqueeText div.
MarqueeText on moon.Item is trying to create marqueeTextWrapper and marqueeText but not able to copy content from the moon.Item.

Fix:
- ItemOverlay.js: When creating beginning and ending block and there is no components block (there is only this.content), we create 'client' and set content.
- Marquee.js: Conditionally return true from _marquee_startAnimation only when we created marqueeText. That's to give a chance to marqueeText in child component start animation.
- Marquee.js: Do not create marqueeTextWrapper and apply animation style when there is components block.

Fix me:
- Needs clean up on initComponents of ItemOverlay. Currently not supporting multiple controls on components block.

DCO-1.1-Signed-Off-By: Kunmyon Choi kunmyon.choi@lge.com